+++ src/StorageKey/IdentityDirectionStorageKey.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKey;
+
+use Heptacom\HeptaConnect\Storage\Base\Contract\IdentityDirectionKeyInterface;
+
+final class IdentityDirectionStorageKey extends AbstractStorageKey implements IdentityDirectionKeyInterface
+{
+}